Move classes from c.s.c.exceptions to c.s.c.rest package
