--- src/main/java/com/streamr/client/exceptions/AmbiguousResultsException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.exceptions;
-
-public class AmbiguousResultsException extends RuntimeException {
-
-    public AmbiguousResultsException(String message) {
-        super(message);
-    }
-
-}
