--- src/main/java/com/streamr/client/exceptions/AuthenticationException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.exceptions;
-
-public class AuthenticationException extends RuntimeException {
-
-    public AuthenticationException(String resourceName) {
-        super("Authentication failed: " + resourceName);
-    }
-
-}
