--- src/main/java/com/streamr/client/exceptions/PermissionDeniedException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.exceptions;
-
-public class PermissionDeniedException extends RuntimeException {
-
-    public PermissionDeniedException(String resourceName) {
-        super("Permission denied: " + resourceName);
-    }
-
-}
