--- src/main/java/com/streamr/client/exceptions/ResourceNotFoundException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.exceptions;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String resourceName) {
-        super("Resource not found: " + resourceName);
-    }
-
-}
