+++ src/main/java/com/streamr/client/rest/AmbiguousResultsException.java
@@ -1,0 +1,7 @@
+package com.streamr.client.rest;
+
+public class AmbiguousResultsException extends RuntimeException {
+  public AmbiguousResultsException(String message) {
+    super(message);
+  }
+}
