+++ src/main/java/com/streamr/client/rest/AuthenticationException.java
@@ -1,0 +1,7 @@
+package com.streamr.client.rest;
+
+public class AuthenticationException extends RuntimeException {
+  public AuthenticationException(String resourceName) {
+    super("Authentication failed: " + resourceName);
+  }
+}
