+++ src/main/java/com/streamr/client/rest/PermissionDeniedException.java
@@ -1,0 +1,8 @@
+package com.streamr.client.rest;
+
+public class PermissionDeniedException extends RuntimeException {
+
+  public PermissionDeniedException(String resourceName) {
+    super("Permission denied: " + resourceName);
+  }
+}
