+++ src/main/java/com/streamr/client/rest/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.streamr.client.rest;
+
+public class ResourceNotFoundException extends RuntimeException {
+  public ResourceNotFoundException(String resourceName) {
+    super("Resource not found: " + resourceName);
+  }
+}
