+++ src/main/java/com/streamr/client/utils/HttpUtils.java
@@ -3,9 +3,9 @@ package com.streamr.client.utils;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
-import com.streamr.client.exceptions.AuthenticationException;
+import com.streamr.client.rest.AuthenticationException;
 import com.streamr.client.exceptions.PermissionDeniedException;
-import com.streamr.client.exceptions.ResourceNotFoundException;
+import com.streamr.client.rest.ResourceNotFoundException;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.net.HttpURLConnection;
